refactor(codenexus): split scratch/timeout helpers out of builtin_provider_evidence.rs
The sentrux architecture gate failed CI (exit 10, "god_files_increased":
builtin_provider_evidence.rs reached 814 loc / 23 functions, crossing the
loc>800 threshold) after the timeout+cleanup commit pushed it over the
limit -- a real, deterministic gate, not a flake. Missed locally because
I checked evidence.codenexus's own status but not diagnosis.hospital's.

Extracted the CodeNexus-lite-specific process/filesystem plumbing that
has nothing to do with the admission/evidence contract itself into a new
crates/code-intel-cli/src/codenexus_scratch.rs: CODENEXUS_LITE_TIMEOUT,
create_codenexus_scratch_dir, ScratchDir, run_with_timeout. Brings
builtin_provider_evidence.rs to 757 lines. codenexus_admission itself,
and the graph/sentrux admission functions, are untouched.

Resynced every toolchain digest chained through the two changed files
(graph-adapt/sentrux-adapt/codenexus-adapt, ast-grep, rg) and
regenerated the E04 retirement packet again (its frozen source set also
includes builtin_provider_evidence.rs).

Verified: cargo fmt/check clean, full suite 3783/3783, PS
atomic-capability-contract and retirement-packets suites both green,
and a local self-scan confirms diagnosis.hospital now passes (was
"architecture gate failure" / god_files_increased before this commit).
