+++ crates/code-intel-cli/tests/capability_exec.rs
@@ -14,7 +14,7 @@ const STRUCTURED_EDIT_DIGEST: &str =
 const REPO_SNAPSHOT_DIGEST: &str =
     "1fa974b28cc2ab92b1ac04e22702b6b126023bcc0ccf1feee14e3d31a833f9dc";
 const CODENEXUS_TOOLCHAIN_DIGESTS: [&str; 5] = [
-    "d72912f1185aefed0fd82081cde7a9b2e48645b2620b47a877c82f3d51266f89",
+    "f13066fdc1a0242006fa634dc530e33b43cff81b82aafdbc3b99a34c05b5d247",
     "645675312135932dfce365a8dfc14e214cec78ee733f248606547b3eaa56edc8",
     "52644a812174988ede91d98ddfec63c6a91f8478277d7bf74c73f106dd0f776b",
     "98ccc64478b2c61bfd7af741ea1f8ee01a88094065c0f025700e8110b525ef26",
