+++ orchestration/integrations.json
@@ -535,7 +535,7 @@
         "id": "provider.graph-builtin.compat",
         "version": "1.0.0",
         "toolchainDigests": [
-              "d72912f1185aefed0fd82081cde7a9b2e48645b2620b47a877c82f3d51266f89",
+              "f13066fdc1a0242006fa634dc530e33b43cff81b82aafdbc3b99a34c05b5d247",
               "59ef3070547d4af845629a160fdf37e8e1b3edbf9883f5ff882b99839e62c82d",
               "68a4f55b76080575721cde3ee62fd746999c6df8e79303b809baecf5d2514338",
               "52644a812174988ede91d98ddfec63c6a91f8478277d7bf74c73f106dd0f776b",
@@ -587,7 +587,7 @@
         "id": "provider.sentrux-builtin.compat",
         "version": "1.0.0",
         "toolchainDigests": [
-              "d72912f1185aefed0fd82081cde7a9b2e48645b2620b47a877c82f3d51266f89",
+              "f13066fdc1a0242006fa634dc530e33b43cff81b82aafdbc3b99a34c05b5d247",
               "29ad124d984f6ab0756bdc7c0b523a84dfada2a655fe0d282ce681165f6bae4f",
               "f69776e6800f917099794a5a28dbdb1117eaf657568dd09e34bf33e4207ddb4c",
               "52644a812174988ede91d98ddfec63c6a91f8478277d7bf74c73f106dd0f776b",
@@ -662,7 +662,7 @@
           "id": "provider.codenexus-builtin.compat",
           "version": "1.0.0",
           "toolchainDigests": [
-                "d72912f1185aefed0fd82081cde7a9b2e48645b2620b47a877c82f3d51266f89",
+                "f13066fdc1a0242006fa634dc530e33b43cff81b82aafdbc3b99a34c05b5d247",
                 "645675312135932dfce365a8dfc14e214cec78ee733f248606547b3eaa56edc8",
                 "52644a812174988ede91d98ddfec63c6a91f8478277d7bf74c73f106dd0f776b",
                 "98ccc64478b2c61bfd7af741ea1f8ee01a88094065c0f025700e8110b525ef26",
