+++ orchestration/internalization/ast-grep.json
@@ -2,15 +2,15 @@
   "schema": "code-intel-internalization-record.v1",
   "id": "internalization.ast-grep-record",
   "projectId": "code-intel-pipeline",
-  "subject": { "name": "ast-grep structural search executable", "kind": "adapted_capability", "source": { "uri": "https://github.com/ast-grep/ast-grep; installed-evidence=ast-grep 0.42.3", "revision": "installed-version:0.42.3; local-native-source-sha256:74e21715c584392f0955e797817f2ef87e3fa9f21045f53c0d4c1ad8f9c70119; local-conformance-sha256:a09a6d358651757d7c490a79ee70b9d40fbec44f033acc4a6b836451b1a9df35" }, "license": { "id": "MIT-UPSTREAM-CLAIM-LOCAL-COPY-MISSING", "obligations": ["do not redistribute or upgrade from this record until the MIT license text and package provenance are retained locally", "preserve preview-only authority, snapshot lease binding, scope and escape guards, generated-content exclusions, and failure semantics exactly"] } },
+  "subject": { "name": "ast-grep structural search executable", "kind": "adapted_capability", "source": { "uri": "https://github.com/ast-grep/ast-grep; installed-evidence=ast-grep 0.42.3", "revision": "installed-version:0.42.3; local-native-source-sha256:74e21715c584392f0955e797817f2ef87e3fa9f21045f53c0d4c1ad8f9c70119; local-conformance-sha256:f22e510efc823eb95abc3fcbeb5aca7c0fe24d348ed65ad6da9650bc52cda70b" }, "license": { "id": "MIT-UPSTREAM-CLAIM-LOCAL-COPY-MISSING", "obligations": ["do not redistribute or upgrade from this record until the MIT license text and package provenance are retained locally", "preserve preview-only authority, snapshot lease binding, scope and escape guards, generated-content exclusions, and failure semantics exactly"] } },
   "adoption": { "rung": "invoke", "ownedBoundary": ["Pipeline owns edit.ast-grep-plan request validation, snapshot lease, path scope and escape guards, artifact contract, and failure mapping", "ast-grep owns pattern parsing, structural matching, and rewrite preview computation; this record does not authorize upgrade, vendoring, repository mutation, or reimplementation"], "necessityEvidence": { "evidenceIds": ["local:registry:edit.ast-grep-plan:optional", "local:i40:installed-ast-grep-0.42.3", "gap:ast-grep:package-provenance"], "checkedAt": 1785024000, "expiresAt": 1792800000 }, "compatibilityEvidence": { "evidenceIds": ["local:i40:preview-only-artifact-contract", "gap:ast-grep:replacement-command-drill"], "checkedAt": 1785024000, "expiresAt": 1792800000 }, "conformanceEvidence": { "evidenceIds": ["local:i40:scope-and-escape-conformance", "local:i40:operation-trace", "local:i41:ci-platform-matrix"], "checkedAt": 1785024000, "expiresAt": 1792800000 } },
   "operationTrace": [
-    { "integrationId": "edit.ast-grep-plan", "operation": "capabilityExec", "command": "target/debug/code-intel.exe capability exec edit.ast-grep-plan --request <request.json|-> --out <staging-dir>", "implementationIdentity": { "providerId": "ast-grep", "implementationId": "edit.ast-grep-plan.compat+ast-grep-0.42.3", "activation": "optional preview-only capability envelope" }, "source": { "path": "crates/code-intel-cli/src/structured_edit.rs", "sha256": "74e21715c584392f0955e797817f2ef87e3fa9f21045f53c0d4c1ad8f9c70119" }, "conformance": { "path": "crates/code-intel-cli/tests/capability_exec.rs", "sha256": "a09a6d358651757d7c490a79ee70b9d40fbec44f033acc4a6b836451b1a9df35", "testName": "structured_edit_plan_is_scope_bound_and_preview_only" } }
+    { "integrationId": "edit.ast-grep-plan", "operation": "capabilityExec", "command": "target/debug/code-intel.exe capability exec edit.ast-grep-plan --request <request.json|-> --out <staging-dir>", "implementationIdentity": { "providerId": "ast-grep", "implementationId": "edit.ast-grep-plan.compat+ast-grep-0.42.3", "activation": "optional preview-only capability envelope" }, "source": { "path": "crates/code-intel-cli/src/structured_edit.rs", "sha256": "74e21715c584392f0955e797817f2ef87e3fa9f21045f53c0d4c1ad8f9c70119" }, "conformance": { "path": "crates/code-intel-cli/tests/capability_exec.rs", "sha256": "f22e510efc823eb95abc3fcbeb5aca7c0fe24d348ed65ad6da9650bc52cda70b", "testName": "structured_edit_plan_is_scope_bound_and_preview_only" } }
   ],
   "economics": { "benefit": { "metric": "registered preview-only structural edit operations with recomputable invocation trace", "value": 1, "unit": "operations" }, "cost": { "metric": "unclosed executable lifecycle gaps", "value": 4, "unit": "gaps" }, "benefitEvidence": { "evidenceIds": ["local:i40:operation-trace", "local:i40:scope-and-escape-conformance"], "checkedAt": 1785024000, "expiresAt": 1792800000 }, "costEvidence": { "evidenceIds": ["gap:ast-grep:package-provenance", "gap:ast-grep:local-license-copy", "gap:ast-grep:replacement-command-drill", "gap:ast-grep:latency-p50-p95-measurement"], "checkedAt": 1785024000, "expiresAt": 1792800000 } },
   "assurance": { "maintenanceEvidence": { "evidenceIds": ["local:i40:pinned-installed-version", "gap:ast-grep:upstream-maintenance-review"], "checkedAt": 1785024000, "expiresAt": 1792800000 }, "securityEvidence": { "evidenceIds": ["local:i40:no-network-read-only-invocation", "local:i41:ci-pinned-artifact-digests", "gap:ast-grep:package-supply-chain-review"], "checkedAt": 1785024000, "expiresAt": 1792800000 } },
   "update": { "policy": "Do not upgrade ast-grep implicitly; before 2026-10-24 retain package provenance/license, rerun the scope and escape conformance test on the CI platform matrix, and refresh the pinned release artifact digests in ci.yml together with this record", "nextCheckAt": 1792800000, "evidence": { "evidenceIds": ["gap:ast-grep:update-review"], "checkedAt": 1785024000, "expiresAt": 1792800000 } },
-  "ownedModifications": [{ "path": "crates/code-intel-cli/src/structured_edit.rs", "description": "Pipeline-owned preview-only planning adapter and snapshot-bound invocation boundary", "evidenceIds": ["local:i40:native-source-sha256:74e21715c584392f0955e797817f2ef87e3fa9f21045f53c0d4c1ad8f9c70119", "local:i40:conformance-sha256:a09a6d358651757d7c490a79ee70b9d40fbec44f033acc4a6b836451b1a9df35"] }],
+  "ownedModifications": [{ "path": "crates/code-intel-cli/src/structured_edit.rs", "description": "Pipeline-owned preview-only planning adapter and snapshot-bound invocation boundary", "evidenceIds": ["local:i40:native-source-sha256:74e21715c584392f0955e797817f2ef87e3fa9f21045f53c0d4c1ad8f9c70119", "local:i40:conformance-sha256:f22e510efc823eb95abc3fcbeb5aca7c0fe24d348ed65ad6da9650bc52cda70b"] }],
   "rollback": { "strategy": "disable the optional edit.ast-grep-plan capability without changing any other artifact contract; no consumer holds authority through it", "evidence": { "evidenceIds": ["local:registry:edit.ast-grep-plan:optional", "gap:ast-grep:replacement-command-drill"], "checkedAt": 1785024000, "expiresAt": 1792800000 } },
   "exit": { "strategy": "replace the executable only behind edit.ast-grep-plan after exact artifact and failure parity", "replacementCriteria": ["alternate command passes pattern, rewrite, scope, escape, generated-content, snapshot-lease, and oversized-output fixtures", "representative latency p50/p95 and cost do not regress beyond the approved budget", "new executable has pinned provenance, license, security, update, rollback, and retirement evidence"], "evidence": { "evidenceIds": ["gap:ast-grep:replacement-command-drill", "gap:ast-grep:latency-p50-p95-measurement"], "checkedAt": 1785024000, "expiresAt": 1792800000 } },
   "retirement": { "status": "candidate", "triggers": ["replacement passes the complete structural edit planning contract", "installed executable identity or package provenance becomes unverifiable", "security or maintenance policy rejects the pinned package"], "evidence": { "evidenceIds": ["local:i40:operation-trace", "gap:ast-grep:replacement-command-drill"], "checkedAt": 1785024000, "expiresAt": 1792800000 } },
