+++ orchestration/internalization/codenexus.json
@@ -13,7 +13,7 @@
   "operationTrace": [
     { "integrationId": "provider.codenexus-adapt", "operation": "adapt", "command": "target/debug/code-intel.exe provider codenexus-adapt --request <native.json|-> --artifact-root <artifact-directory> --evaluated-at <unix-seconds> --max-age-seconds <seconds>", "implementationIdentity": { "providerId": "codenexus.full", "implementationId": "codenexus.service.v1", "activation": "primary" }, "source": { "path": "crates/code-intel-cli/src/codenexus_adapter.rs", "sha256": "645675312135932dfce365a8dfc14e214cec78ee733f248606547b3eaa56edc8" }, "conformance": { "path": "crates/code-intel-cli/tests/codenexus_adapter.rs", "sha256": "4b2ea42c33aa9c180df550278d6e74501deb9f3f2c6133f7c2def850e468bdc5", "testName": "production_route_runs_full_lite_and_unavailable_through_a04" } },
     { "integrationId": "provider.codenexus-adapt", "operation": "facade", "command": "legacy/run-code-intel.ps1 -CodeNexusAdapterRequest <native.json> -CodeNexusAdapterArtifactRoot <artifact-directory> -CodeNexusAdapterEvaluatedAt <unix-seconds> -CodeNexusAdapterMaxAgeSeconds <seconds>", "implementationIdentity": { "providerId": "codenexus.lite-compat", "implementationId": "invoke-codenexus-lite.ps1", "activation": "explicit_fallback" }, "source": { "path": "crates/code-intel-cli/src/codenexus_adapter.rs", "sha256": "645675312135932dfce365a8dfc14e214cec78ee733f248606547b3eaa56edc8" }, "conformance": { "path": "crates/code-intel-cli/tests/codenexus_adapter.rs", "sha256": "4b2ea42c33aa9c180df550278d6e74501deb9f3f2c6133f7c2def850e468bdc5", "testName": "production_registry_facade_and_route_schema_are_declared" } },
-    { "integrationId": "provider.codenexus-adapt", "operation": "capabilityExec", "command": "target/debug/code-intel.exe capability exec provider.codenexus-adapt --request <request.json|-> --out <staging-dir> --artifact-root <run-root>", "implementationIdentity": { "providerId": "codenexus.lite-compat", "implementationId": "invoke-codenexus-lite.ps1", "activation": "legacy_rollback" }, "source": { "path": "crates/code-intel-cli/src/builtin_provider_evidence.rs", "sha256": "d72912f1185aefed0fd82081cde7a9b2e48645b2620b47a877c82f3d51266f89" }, "conformance": { "path": "crates/code-intel-cli/tests/capability_exec.rs", "sha256": "a09a6d358651757d7c490a79ee70b9d40fbec44f033acc4a6b836451b1a9df35", "testName": "codenexus_builtin_compat_dispatches_through_provider_codenexus_adapt" } },
+    { "integrationId": "provider.codenexus-adapt", "operation": "capabilityExec", "command": "target/debug/code-intel.exe capability exec provider.codenexus-adapt --request <request.json|-> --out <staging-dir> --artifact-root <run-root>", "implementationIdentity": { "providerId": "codenexus.lite-compat", "implementationId": "invoke-codenexus-lite.ps1", "activation": "legacy_rollback" }, "source": { "path": "crates/code-intel-cli/src/builtin_provider_evidence.rs", "sha256": "f13066fdc1a0242006fa634dc530e33b43cff81b82aafdbc3b99a34c05b5d247" }, "conformance": { "path": "crates/code-intel-cli/tests/capability_exec.rs", "sha256": "f22e510efc823eb95abc3fcbeb5aca7c0fe24d348ed65ad6da9650bc52cda70b", "testName": "codenexus_builtin_compat_dispatches_through_provider_codenexus_adapt" } },
     { "integrationId": "runtime.code-nexus-lite", "operation": "compat", "command": "pwsh -NoProfile -File \"$env:CODE_INTEL_HOME\\legacy/Invoke-CodeNexusLite.ps1\" -RepoPath '<repo-path>'", "implementationIdentity": { "providerId": "codenexus.lite-compat", "implementationId": "invoke-codenexus-lite.ps1", "activation": "explicit_fallback" }, "source": { "path": "legacy/Invoke-CodeNexusLite.ps1", "sha256": "cdd5c6d0fe940d2756c45a51095c275b13ebe64347914b581641693e1288ca56" }, "conformance": { "path": "legacy/scripts/tests/test-codenexus-adapter-contract.ps1", "sha256": "a835382e48223e35d680f29d27e05fc12c9d92306e1bafafc7add1135f26b1a8", "testName": "Invoke-LiteScriptEndToEnd" } },
     { "integrationId": "localization.codenexus-lite", "operation": "compat", "command": "pwsh -NoProfile -File \"$env:CODE_INTEL_HOME\\legacy/Invoke-CodeNexusLite.ps1\" -RepoPath '<repo-path>'", "implementationIdentity": { "providerId": "codenexus.lite-compat", "implementationId": "invoke-codenexus-lite.ps1", "activation": "legacy_rollback" }, "source": { "path": "legacy/Invoke-CodeNexusLite.ps1", "sha256": "cdd5c6d0fe940d2756c45a51095c275b13ebe64347914b581641693e1288ca56" }, "conformance": { "path": "legacy/scripts/tests/test-codenexus-adapter-contract.ps1", "sha256": "a835382e48223e35d680f29d27e05fc12c9d92306e1bafafc7add1135f26b1a8", "testName": "Invoke-LiteScriptEndToEnd" } }
   ],
