+++ orchestration/internalization/graph.json
@@ -13,7 +13,7 @@
   "operationTrace": [
     { "integrationId": "provider.graph-adapt", "operation": "adapt", "command": "target/debug/code-intel.exe provider graph-adapt --request <native.json|-> --artifact-root <artifact-directory> --evaluated-at <unix-seconds> --max-age-seconds <seconds>", "implementationIdentity": { "providerId": "code-intel.graph", "implementationId": "graph-adapter.v1", "activation": "primary" }, "source": { "path": "crates/code-intel-cli/src/graph_adapter.rs", "sha256": "68a4f55b76080575721cde3ee62fd746999c6df8e79303b809baecf5d2514338" }, "conformance": { "path": "crates/code-intel-cli/tests/graph_adapter.rs", "sha256": "024625a11f874ccc93185a82e3285845150ebea301909bd23e3dfa9becd76f32", "testName": "public_route_usage_registry_facade_and_schemas_are_real" } },
     { "integrationId": "provider.graph-adapt", "operation": "facade", "command": "legacy/run-code-intel.ps1 -GraphAdapterRequest <native.json> -GraphAdapterArtifactRoot <artifact-directory> -GraphAdapterEvaluatedAt <unix-seconds> -GraphAdapterMaxAgeSeconds <seconds>", "implementationIdentity": { "providerId": "code-intel.graph", "implementationId": "graph-adapter.v1", "activation": "compatibility_facade" }, "source": { "path": "crates/code-intel-cli/src/graph_adapter.rs", "sha256": "68a4f55b76080575721cde3ee62fd746999c6df8e79303b809baecf5d2514338" }, "conformance": { "path": "crates/code-intel-cli/tests/graph_adapter.rs", "sha256": "024625a11f874ccc93185a82e3285845150ebea301909bd23e3dfa9becd76f32", "testName": "public_route_usage_registry_facade_and_schemas_are_real" } },
-    { "integrationId": "provider.graph-adapt", "operation": "capabilityExec", "command": "target/debug/code-intel.exe capability exec provider.graph-adapt --request <request.json|-> --out <staging-dir> --artifact-root <run-root>", "implementationIdentity": { "providerId": "code-intel.graph", "implementationId": "provider.graph-builtin.compat", "activation": "required production capability envelope" }, "source": { "path": "crates/code-intel-cli/src/builtin_provider_evidence.rs", "sha256": "d72912f1185aefed0fd82081cde7a9b2e48645b2620b47a877c82f3d51266f89" }, "conformance": { "path": "crates/code-intel-cli/tests/dag_run.rs", "sha256": "14a4af8625933b9f12494b8a2b53e940dea2246d1501f027b718ffa1fe75caa0", "testName": "production_dag_output_commits_and_enters_the_authoritative_index" } },
+    { "integrationId": "provider.graph-adapt", "operation": "capabilityExec", "command": "target/debug/code-intel.exe capability exec provider.graph-adapt --request <request.json|-> --out <staging-dir> --artifact-root <run-root>", "implementationIdentity": { "providerId": "code-intel.graph", "implementationId": "provider.graph-builtin.compat", "activation": "required production capability envelope" }, "source": { "path": "crates/code-intel-cli/src/builtin_provider_evidence.rs", "sha256": "f13066fdc1a0242006fa634dc530e33b43cff81b82aafdbc3b99a34c05b5d247" }, "conformance": { "path": "crates/code-intel-cli/tests/dag_run.rs", "sha256": "14a4af8625933b9f12494b8a2b53e940dea2246d1501f027b718ffa1fe75caa0", "testName": "production_dag_output_commits_and_enters_the_authoritative_index" } },
     { "integrationId": "graph.code-intel-understand", "operation": "refresh", "command": "target/debug/code-intel.exe graph --repo <repo-path> --language zh --write --json", "implementationIdentity": { "providerId": "code-intel.graph", "implementationId": "internal-rust-graph", "activation": "primary" }, "source": { "path": "crates/code-intel-cli/src/graph/mod.rs", "sha256": "59ef3070547d4af845629a160fdf37e8e1b3edbf9883f5ff882b99839e62c82d" }, "conformance": { "path": "crates/code-intel-cli/tests/graph_adapter.rs", "sha256": "024625a11f874ccc93185a82e3285845150ebea301909bd23e3dfa9becd76f32", "testName": "internal_and_external_current_outputs_share_one_port_and_provenance_schema" } },
     { "integrationId": "graph.code-intel-understand", "operation": "refreshFull", "command": "target/debug/code-intel.exe graph --repo <repo-path> --language zh --full --write --json", "implementationIdentity": { "providerId": "code-intel.graph", "implementationId": "internal-rust-graph", "activation": "primary" }, "source": { "path": "crates/code-intel-cli/src/graph/mod.rs", "sha256": "59ef3070547d4af845629a160fdf37e8e1b3edbf9883f5ff882b99839e62c82d" }, "conformance": { "path": "crates/code-intel-cli/tests/graph_adapter.rs", "sha256": "024625a11f874ccc93185a82e3285845150ebea301909bd23e3dfa9becd76f32", "testName": "internal_and_external_current_outputs_share_one_port_and_provenance_schema" } },
     { "integrationId": "graph.understand-external", "operation": "refresh", "command": "/understand <repo-path> --language zh", "implementationIdentity": { "providerId": "understand-anything", "implementationId": "unverified-upstream", "activation": "explicit_fallback" }, "source": { "path": "docs/graph-provider-adapter.md", "sha256": "21a6633e164aa5aab122e1d2de36fad47866611abe930b64d03bb67633408095" }, "conformance": { "path": "crates/code-intel-cli/tests/graph_adapter.rs", "sha256": "024625a11f874ccc93185a82e3285845150ebea301909bd23e3dfa9becd76f32", "testName": "fallback_identity_and_payload_identity_cannot_be_relabelled" } },
