+++ orchestration/retirements/e04-codenexus-direct/compatibility-retirement-ticket.json
@@ -1,1 +1,1 @@
-{"schema":"code-intel-compatibility-retirement-ticket-template.v1","snapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc","ticketId":"ticket-e04-retire-codenexus-direct-branch","retirementId":"retire-codenexus-direct-branch","legacyBranch":{"capabilityId":"localization.codenexus-lite","branchId":"run-code-intel.codenexus-lite.direct","callPath":"run-code-intel.ps1::run-code-intel.codenexus-lite.direct"},"replacement":{"capabilityId":"provider.codenexus-adapt","dependencies":["repository.survival-scan"]},"affectedFiles":["run-code-intel.ps1"],"evidence":{"golden":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-evidence.v1","type":"compatibility.retirement-evidence","path":"evidence/golden-parity.json","sha256":"a51196a1af936a6e1fe1ceae5cee28b6d98d7481d9856c950f676a140e39f699","consumedSnapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc"},"contract":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-evidence.v1","type":"compatibility.retirement-evidence","path":"evidence/contract-parity.json","sha256":"26d1e3b48832c41448e033ea3782dbcedffe2b84889feda0830fb4bd4bba24f3","consumedSnapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc"},"effects":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-evidence.v1","type":"compatibility.retirement-evidence","path":"evidence/effect-parity.json","sha256":"f1344c43cac53f870117897eb056d054dc00609c489c35cf4734a659c13a2fb4","consumedSnapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc"},"usage":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-evidence.v1","type":"compatibility.retirement-evidence","path":"evidence/usage-observation.json","sha256":"171e1462e3295256822595afb4a94483fa8e360b3dd86585d23e8fcaf833a683","consumedSnapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc"},"rollbackRehearsal":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-evidence.v1","type":"compatibility.retirement-evidence","path":"evidence/rollback-execution.json","sha256":"4e0c4ea661c3238d56d77a140b35f83f80d0ef8187b5d42905d71cb059d58725","consumedSnapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc"},"deletionDiff":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-deletion-diff.v1","type":"compatibility.retirement-deletion-diff","path":"compatibility-retirement-deletion-diff.json","sha256":"26aee5622cd2bb282849910e7ec8324b0283136f599d14b5a30a35989c11baa7","consumedSnapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc"}},"source":{"retirementDecision":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-decision.v1","type":"compatibility.retirement-decision","path":"gate-out/compatibility-retirement-decision.json","sha256":"32f7b1f714ad8c5885543e44ce8c3c4725c08e8dc1c57ce3675ddcb8d1aadf20","consumedSnapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc"},"retirementManifest":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-manifest.v1","type":"compatibility.retirement-manifest","path":"compatibility-retirement-manifest.json","sha256":"fe70bae257eaca64403ac1f87a048a91ef1f66d3ed73e77eeb1d8c181dea2f4b","consumedSnapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc"}},"owner":"executor-codenexus-retirement","verifier":"independent-verifier-required","observationExpiry":1788680980,"status":"draft","authorityBoundary":"template_only_no_approval_or_deletion_authority"}
+{"schema":"code-intel-compatibility-retirement-ticket-template.v1","snapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04","ticketId":"ticket-e04-retire-codenexus-direct-branch","retirementId":"retire-codenexus-direct-branch","legacyBranch":{"capabilityId":"localization.codenexus-lite","branchId":"run-code-intel.codenexus-lite.direct","callPath":"run-code-intel.ps1::run-code-intel.codenexus-lite.direct"},"replacement":{"capabilityId":"provider.codenexus-adapt","dependencies":["repository.survival-scan"]},"affectedFiles":["run-code-intel.ps1"],"evidence":{"golden":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-evidence.v1","type":"compatibility.retirement-evidence","path":"evidence/golden-parity.json","sha256":"7ba2b3b680a50de723beacb28e33e854bcb9b6be2a8db0bb30ddf7f88dcf6b0f","consumedSnapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04"},"contract":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-evidence.v1","type":"compatibility.retirement-evidence","path":"evidence/contract-parity.json","sha256":"ffaac33bef5cfaf8ab96b864b883aad33e1619f0fda79145109d7977c6752a21","consumedSnapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04"},"effects":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-evidence.v1","type":"compatibility.retirement-evidence","path":"evidence/effect-parity.json","sha256":"5eafc12d8cb95874171f4cbe4c25ee78f2a3a5b0e13d4b5df248f14a9a6facd0","consumedSnapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04"},"usage":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-evidence.v1","type":"compatibility.retirement-evidence","path":"evidence/usage-observation.json","sha256":"88da7dfc0f3b7a3685bf8e96c9c29bf6a5c3e318649009c7e5ca7ccd3f38a6a4","consumedSnapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04"},"rollbackRehearsal":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-evidence.v1","type":"compatibility.retirement-evidence","path":"evidence/rollback-execution.json","sha256":"693e0b394ffd1028ac01da409012fb99bf019e33b8e7c55241295507be05d00c","consumedSnapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04"},"deletionDiff":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-deletion-diff.v1","type":"compatibility.retirement-deletion-diff","path":"compatibility-retirement-deletion-diff.json","sha256":"85ebe6bb914e24ef3627c0eb29506a3f007a1e1aed67cf0d90b92492de7a6974","consumedSnapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04"}},"source":{"retirementDecision":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-decision.v1","type":"compatibility.retirement-decision","path":"gate-out/compatibility-retirement-decision.json","sha256":"4ef3693ee2588f06a6c19213a393fe92f114888652d97d1a126d67ed8c742c16","consumedSnapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04"},"retirementManifest":{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-manifest.v1","type":"compatibility.retirement-manifest","path":"compatibility-retirement-manifest.json","sha256":"3cd1ba62c1d5a4ec94fdc7ba0d33078a3cd1575ef88b4b43cf353959bd873fc6","consumedSnapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04"}},"owner":"executor-codenexus-retirement","verifier":"independent-verifier-required","observationExpiry":1788683467,"status":"draft","authorityBoundary":"template_only_no_approval_or_deletion_authority"}
