+++ orchestration/retirements/e04-codenexus-direct/e01-request.json
@@ -1,1 +1,1 @@
-{"schema":"code-intel-capability-request.v1","capability":"compatibility.retirement-ticket-template","contractVersion":1,"implementation":{"id":"compatibility.retirement-ticket-template.compat","version":"1.0.0","toolchainDigests":["a6da84444e815dd94c15887190c7d6bd48c7bca5975fb5d493eb77ab39f27d80","255483f15ef2cdfe0b759d84ef64bc20ec5236716ea6668642c6806e834ab437","295eb1ce67760638a81136febf727285f0feb4692a228df65ac75316b4a566c5"]},"snapshot":{"identity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc","repoIdentity":"content-v1:cccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc","head":"unversioned","workingTreePolicy":"explicit_overlay","scope":["."],"inputDigest":"dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd"},"options":{"evaluatedAt":1786088980},"inputs":[{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-ticket-template.v1","type":"compatibility.retirement-ticket-template","path":"compatibility-retirement-ticket.json","sha256":"2ce347c05d76f0b12d1899a790a5d63f055d01dae9c500220cf9fe8974aa5b6d","consumedSnapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc"},{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-manifest.v1","type":"compatibility.retirement-manifest","path":"compatibility-retirement-manifest.json","sha256":"fe70bae257eaca64403ac1f87a048a91ef1f66d3ed73e77eeb1d8c181dea2f4b","consumedSnapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc"},{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-decision.v1","type":"compatibility.retirement-decision","path":"gate-out/compatibility-retirement-decision.json","sha256":"32f7b1f714ad8c5885543e44ce8c3c4725c08e8dc1c57ce3675ddcb8d1aadf20","consumedSnapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc"},{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-deletion-diff.v1","type":"compatibility.retirement-deletion-diff","path":"compatibility-retirement-deletion-diff.json","sha256":"26aee5622cd2bb282849910e7ec8324b0283136f599d14b5a30a35989c11baa7","consumedSnapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc"}],"effectPolicy":{"allowedEffects":["local_write"]}}
+{"schema":"code-intel-capability-request.v1","capability":"compatibility.retirement-ticket-template","contractVersion":1,"implementation":{"id":"compatibility.retirement-ticket-template.compat","version":"1.0.0","toolchainDigests":["a6da84444e815dd94c15887190c7d6bd48c7bca5975fb5d493eb77ab39f27d80","255483f15ef2cdfe0b759d84ef64bc20ec5236716ea6668642c6806e834ab437","295eb1ce67760638a81136febf727285f0feb4692a228df65ac75316b4a566c5"]},"snapshot":{"identity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04","repoIdentity":"content-v1:cccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc","head":"unversioned","workingTreePolicy":"explicit_overlay","scope":["."],"inputDigest":"dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd"},"options":{"evaluatedAt":1786091467},"inputs":[{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-ticket-template.v1","type":"compatibility.retirement-ticket-template","path":"compatibility-retirement-ticket.json","sha256":"ea37bd7c3b16942a31fd597c4c8c12061edf60dbd9b8631298893aff89a85a4e","consumedSnapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04"},{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-manifest.v1","type":"compatibility.retirement-manifest","path":"compatibility-retirement-manifest.json","sha256":"3cd1ba62c1d5a4ec94fdc7ba0d33078a3cd1575ef88b4b43cf353959bd873fc6","consumedSnapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04"},{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-decision.v1","type":"compatibility.retirement-decision","path":"gate-out/compatibility-retirement-decision.json","sha256":"4ef3693ee2588f06a6c19213a393fe92f114888652d97d1a126d67ed8c742c16","consumedSnapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04"},{"schema":"code-intel-artifact-ref.v1","artifactSchema":"code-intel-compatibility-retirement-deletion-diff.v1","type":"compatibility.retirement-deletion-diff","path":"compatibility-retirement-deletion-diff.json","sha256":"85ebe6bb914e24ef3627c0eb29506a3f007a1e1aed67cf0d90b92492de7a6974","consumedSnapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04"}],"effectPolicy":{"allowedEffects":["local_write"]}}
