+++ orchestration/retirements/e04-codenexus-direct/e01-stderr.txt
@@ -1,2 +1,2 @@
-{"artifacts":[],"cache":{"hit":false,"key":null},"capability":"compatibility.retirement-ticket-template","declaredEffects":["local_write"],"determinism":"deterministic","diagnostics":["ticket requires an approved E00 decision"],"domainVerdict":"unknown","exitCode":65,"implementation":{"id":"compatibility.retirement-ticket-template.compat","toolchainDigests":["a6da84444e815dd94c15887190c7d6bd48c7bca5975fb5d493eb77ab39f27d80","255483f15ef2cdfe0b759d84ef64bc20ec5236716ea6668642c6806e834ab437","295eb1ce67760638a81136febf727285f0feb4692a228df65ac75316b4a566c5"],"version":"1.0.0"},"observedEffects":[],"provenance":{"attemptId":"capability-1786089005-13804","generatedAt":"2026-08-07T07:50:05Z"},"schema":"code-intel-capability-result.v1","snapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc","status":"failed","verdict":"unknown"}
+{"artifacts":[],"cache":{"hit":false,"key":null},"capability":"compatibility.retirement-ticket-template","declaredEffects":["local_write"],"determinism":"deterministic","diagnostics":["ticket requires an approved E00 decision"],"domainVerdict":"unknown","exitCode":65,"implementation":{"id":"compatibility.retirement-ticket-template.compat","toolchainDigests":["a6da84444e815dd94c15887190c7d6bd48c7bca5975fb5d493eb77ab39f27d80","255483f15ef2cdfe0b759d84ef64bc20ec5236716ea6668642c6806e834ab437","295eb1ce67760638a81136febf727285f0feb4692a228df65ac75316b4a566c5"],"version":"1.0.0"},"observedEffects":[],"provenance":{"attemptId":"capability-1786091479-19320","generatedAt":"2026-08-07T08:31:19Z"},"schema":"code-intel-capability-result.v1","snapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04","status":"failed","verdict":"unknown"}
 ticket requires an approved E00 decision
