+++ orchestration/retirements/e04-codenexus-direct/evidence/c00-necessity.json
@@ -1,1 +1,1 @@
-{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc","id":"e04.c00-necessity","evidenceClass":"c00_necessity","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"passed","decision":"admit","changeId":"retire-codenexus-direct-branch","necessityTraceSha256":"190b7c08bbb12bec70610c96de9f83119fdce8df960ce0eeac420c440c9b7cd0","rationale":"remove facade-owned provider process/storage coupling only after B04/B05 route replacement"}}
+{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04","id":"e04.c00-necessity","evidenceClass":"c00_necessity","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"passed","decision":"admit","changeId":"retire-codenexus-direct-branch","necessityTraceSha256":"190b7c08bbb12bec70610c96de9f83119fdce8df960ce0eeac420c440c9b7cd0","rationale":"remove facade-owned provider process/storage coupling only after B04/B05 route replacement"}}
