+++ orchestration/retirements/e04-codenexus-direct/evidence/compatibility-window.json
@@ -1,1 +1,1 @@
-{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc","id":"e04.compatibility-window","evidenceClass":"compatibility_window","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"blocked","startedAt":1786088980,"observedThrough":1786088980,"minimumDays":30,"checkedAt":1786088980,"expiresAt":1788680980,"blocker":"no completed 30-day full/lite/unavailable production observation window"}}
+{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04","id":"e04.compatibility-window","evidenceClass":"compatibility_window","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"blocked","startedAt":1786091467,"observedThrough":1786091467,"minimumDays":30,"checkedAt":1786091467,"expiresAt":1788683467,"blocker":"no completed 30-day full/lite/unavailable production observation window"}}
