+++ orchestration/retirements/e04-codenexus-direct/evidence/contract-parity.json
@@ -1,1 +1,1 @@
-{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc","id":"e04.contract-parity","evidenceClass":"contract_parity","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"passed","assertionCount":6,"commands":["cargo test -q -p code-intel --test codenexus_adapter <test-name> -- --exact","cargo test -q -p code-intel --test survival_scan <test-name> -- --exact"],"executedB04TestCount":10,"executedB05TestCount":7,"unavailableRoute":"repository.survival-scan","structuralVerdictWhenUnavailable":"unknown","forbiddenStructuralClaims":["architecture","dependency graph","impact analysis","complete call graph"]}}
+{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04","id":"e04.contract-parity","evidenceClass":"contract_parity","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"passed","assertionCount":6,"commands":["cargo test -q -p code-intel --test codenexus_adapter <test-name> -- --exact","cargo test -q -p code-intel --test survival_scan <test-name> -- --exact"],"executedB04TestCount":10,"executedB05TestCount":7,"unavailableRoute":"repository.survival-scan","structuralVerdictWhenUnavailable":"unknown","forbiddenStructuralClaims":["architecture","dependency graph","impact analysis","complete call graph"]}}
