+++ orchestration/retirements/e04-codenexus-direct/evidence/dependency-b05.json
@@ -1,1 +1,1 @@
-{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc","id":"e04.dependency-b05","evidenceClass":"dependency_approval","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"passed","dependencyId":"repository.survival-scan","status":"approved","reviewer":"e04-author-test"}}
+{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04","id":"e04.dependency-b05","evidenceClass":"dependency_approval","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"passed","dependencyId":"repository.survival-scan","status":"approved","reviewer":"e04-author-test"}}
