+++ orchestration/retirements/e04-codenexus-direct/evidence/effect-parity.json
@@ -1,1 +1,1 @@
-{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc","id":"e04.effect-parity","evidenceClass":"effect_parity","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"passed","assertionCount":5,"processOwnership":"provider","storageOwnership":"provider","transport":"artifact_ref","fullEffects":["network_provider","read_provider_artifact"],"liteEffects":["read_repository","read_git_history","read_sentrux_artifacts","write_compatibility_artifact"],"fallbackEffects":["repo_read","local_write"],"effectSetsRemainModeSpecific":true,"noSecretOrStoragePathProjection":true}}
+{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04","id":"e04.effect-parity","evidenceClass":"effect_parity","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"passed","assertionCount":5,"processOwnership":"provider","storageOwnership":"provider","transport":"artifact_ref","fullEffects":["network_provider","read_provider_artifact"],"liteEffects":["read_repository","read_git_history","read_sentrux_artifacts","write_compatibility_artifact"],"fallbackEffects":["repo_read","local_write"],"effectSetsRemainModeSpecific":true,"noSecretOrStoragePathProjection":true}}
