+++ orchestration/retirements/e04-codenexus-direct/evidence/golden-parity.json
@@ -1,1 +1,1 @@
-{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc","id":"e04.golden-parity","evidenceClass":"golden_parity","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"passed","assertionCount":3,"command":"cargo test -q -p code-intel --test codenexus_adapter <test-name> -- --exact","executedTestCount":10,"testFunctions":["full_and_lite_share_one_port_shape_but_keep_distinct_provenance","unavailable_is_partial_provider_unavailable_and_never_fabricates_facts","snapshot_mismatch_and_stale_observation_fail_closed_in_a04","lite_is_only_explicit_fallback_or_rollback_and_full_is_primary","adapter_declares_effects_rejects_storage_coupling_and_drops_unknown_secrets","admitted_payload_identity_cannot_be_relabelled","port_schema_and_boundary_document_are_closed_and_real","production_route_runs_full_lite_and_unavailable_through_a04","production_route_rejects_secret_fields_wrong_snapshot_and_bad_usage","production_registry_facade_and_route_schema_are_declared"],"fixtures":["full-current","lite-current","unavailable"],"portShapeShared":true,"provenanceDistinct":true}}
+{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04","id":"e04.golden-parity","evidenceClass":"golden_parity","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"passed","assertionCount":3,"command":"cargo test -q -p code-intel --test codenexus_adapter <test-name> -- --exact","executedTestCount":10,"testFunctions":["full_and_lite_share_one_port_shape_but_keep_distinct_provenance","unavailable_is_partial_provider_unavailable_and_never_fabricates_facts","snapshot_mismatch_and_stale_observation_fail_closed_in_a04","lite_is_only_explicit_fallback_or_rollback_and_full_is_primary","adapter_declares_effects_rejects_storage_coupling_and_drops_unknown_secrets","admitted_payload_identity_cannot_be_relabelled","port_schema_and_boundary_document_are_closed_and_real","production_route_runs_full_lite_and_unavailable_through_a04","production_route_rejects_secret_fields_wrong_snapshot_and_bad_usage","production_registry_facade_and_route_schema_are_declared"],"fixtures":["full-current","lite-current","unavailable"],"portShapeShared":true,"provenanceDistinct":true}}
