+++ orchestration/retirements/e04-codenexus-direct/evidence/independent-approval.json
@@ -1,1 +1,1 @@
-{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc","id":"e04.independent-approval","evidenceClass":"independent_approval","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"blocked","approved":false,"authorIndependent":false,"subjectSha256":"0000000000000000000000000000000000000000000000000000000000000000","reviewer":"independent-verifier-required","authorityEvent":{},"blocker":"no independent repository-governed E00 approval exists"}}
+{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04","id":"e04.independent-approval","evidenceClass":"independent_approval","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"blocked","approved":false,"authorIndependent":false,"subjectSha256":"0000000000000000000000000000000000000000000000000000000000000000","reviewer":"independent-verifier-required","authorityEvent":{},"blocker":"no independent repository-governed E00 approval exists"}}
