+++ orchestration/retirements/e04-codenexus-direct/evidence/replacement-atom.json
@@ -1,1 +1,1 @@
-{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc","id":"e04.replacement-atom","evidenceClass":"replacement_atom","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"blocked","status":"pending_facade_route","capability":"provider.codenexus-adapt","b04FacadeAvailable":true,"b05FacadeAvailable":true,"liveDirectBranchCount":1,"blocker":"normal production path still invokes Invoke-CodeNexusLite.ps1 directly instead of composing B04 and B05"}}
+{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04","id":"e04.replacement-atom","evidenceClass":"replacement_atom","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"blocked","status":"pending_facade_route","capability":"provider.codenexus-adapt","b04FacadeAvailable":true,"b05FacadeAvailable":true,"liveDirectBranchCount":1,"blocker":"normal production path still invokes Invoke-CodeNexusLite.ps1 directly instead of composing B04 and B05"}}
