+++ orchestration/retirements/e04-codenexus-direct/evidence/rollback-execution.json
@@ -1,1 +1,1 @@
-{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc","id":"e04.rollback-execution","evidenceClass":"rollback_execution","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"passed","command":"pwsh -NoProfile -File tools/compatibility/Restore-CodeNexusDirectBranch.ps1 -RehearsalRoot <packet-root>/rollback-rehearsal","executedAt":1786088980,"exitCode":0,"target":"rollback-rehearsal/run-code-intel.ps1","exactReplay":true,"unrelatedBranchesChanged":false}}
+{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04","id":"e04.rollback-execution","evidenceClass":"rollback_execution","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"passed","command":"pwsh -NoProfile -File tools/compatibility/Restore-CodeNexusDirectBranch.ps1 -RehearsalRoot <packet-root>/rollback-rehearsal","executedAt":1786091467,"exitCode":0,"target":"rollback-rehearsal/run-code-intel.ps1","exactReplay":true,"unrelatedBranchesChanged":false}}
