+++ orchestration/retirements/e04-codenexus-direct/evidence/usage-observation.json
@@ -1,1 +1,1 @@
-{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc","id":"e04.usage-observation","evidenceClass":"usage_observation","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"blocked","startedAt":1786088980,"endedAt":1786088980,"totalInvocations":0,"legacyInvocations":0,"replacementInvocations":0,"blocker":"no production route/usage observation exists"}}
+{"schema":"code-intel-compatibility-retirement-evidence.v1","snapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04","id":"e04.usage-observation","evidenceClass":"usage_observation","retirementId":"retire-codenexus-direct-branch","legacyBranchId":"run-code-intel.codenexus-lite.direct","replacementCapabilityId":"provider.codenexus-adapt","details":{"outcome":"blocked","startedAt":1786091467,"endedAt":1786091467,"totalInvocations":0,"legacyInvocations":0,"replacementInvocations":0,"blocker":"no production route/usage observation exists"}}
