+++ orchestration/retirements/e04-codenexus-direct/gate-out/compatibility-retirement-decision.json
@@ -1,1 +1,1 @@
-{"approvalSubjectSha256":"472f0da62f494ad47aff6893e9f7e9fa299ecea3b2ba677b8df7698c584a2a3e","authorityBoundary":"approval_only_no_deletion_authority","blockers":["unproven_compatibility_window","unproven_independent_approval","unproven_replacement_atom","unproven_usage_observation"],"decision":"blocked","gainLedgerProjection":{"evidenceCount":11,"gain":"Retire legacy branch run-code-intel.codenexus-lite.direct only through a separate deletion ticket","id":"retirement:retire-codenexus-direct-branch","status":"blocked"},"legacyBranch":{"affectedFiles":["run-code-intel.ps1"],"branchId":"run-code-intel.codenexus-lite.direct","callPath":"run-code-intel.ps1::run-code-intel.codenexus-lite.direct","capabilityId":"localization.codenexus-lite","owner":"executor-codenexus-retirement","registryParticipantId":"localization.codenexus-lite"},"replacement":{"atomEvidence":{"artifactSchema":"code-intel-compatibility-retirement-evidence.v1","consumedSnapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc","path":"evidence/replacement-atom.json","schema":"code-intel-artifact-ref.v1","sha256":"d8907f489198f7e6cd58c641a8f3edaee99a97f299c3efc9f2fb1dd1a533617c","type":"compatibility.retirement-evidence"},"capabilityId":"provider.codenexus-adapt","dependencies":["repository.survival-scan"],"implementationId":"provider.codenexus-adapt"},"retirementId":"retire-codenexus-direct-branch","schema":"code-intel-compatibility-retirement-decision.v1","snapshotIdentity":"75e5583763798835e4170154e504b399c9ce5bd4f28147c3fb3baddafadd17fc"}
+{"approvalSubjectSha256":"29d823c7aba0597da8c7fb71bf8b6f5b045c69281e1bf08138ba8e0334504831","authorityBoundary":"approval_only_no_deletion_authority","blockers":["unproven_compatibility_window","unproven_independent_approval","unproven_replacement_atom","unproven_usage_observation"],"decision":"blocked","gainLedgerProjection":{"evidenceCount":11,"gain":"Retire legacy branch run-code-intel.codenexus-lite.direct only through a separate deletion ticket","id":"retirement:retire-codenexus-direct-branch","status":"blocked"},"legacyBranch":{"affectedFiles":["run-code-intel.ps1"],"branchId":"run-code-intel.codenexus-lite.direct","callPath":"run-code-intel.ps1::run-code-intel.codenexus-lite.direct","capabilityId":"localization.codenexus-lite","owner":"executor-codenexus-retirement","registryParticipantId":"localization.codenexus-lite"},"replacement":{"atomEvidence":{"artifactSchema":"code-intel-compatibility-retirement-evidence.v1","consumedSnapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04","path":"evidence/replacement-atom.json","schema":"code-intel-artifact-ref.v1","sha256":"80ea57ffacc4472dead33f869c613473366f4d285d1f39d18f4777a0c0c5c1ba","type":"compatibility.retirement-evidence"},"capabilityId":"provider.codenexus-adapt","dependencies":["repository.survival-scan"],"implementationId":"provider.codenexus-adapt"},"retirementId":"retire-codenexus-direct-branch","schema":"code-intel-compatibility-retirement-decision.v1","snapshotIdentity":"01d654c63516a476291edae15ec40fc7e416fbb5cf669d4a08f5cd82c31dbb04"}
